#6280 Use Duration for timeouts (#7379)

+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/network/FakeConnection.java
@@ -77,7 +77,7 @@ public class FakeConnection implements Connection {
     }
 
     @Override
-    public void setTimeout(final int timeoutMillis) {
+    public void setTimeout(final long timeoutMillis) {
         throw new UnsupportedOperationException("N/A");
     }
